% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsaav-package.R
\docType{data}
\name{airport_wx}
\alias{airport_wx}
\title{Airport weather data at Narsarsuaq, Greenland}
\format{An object of class \code{data.frame} with 6498 rows and 24 columns.}
\usage{
airport_wx
}
\description{
A 24 variable 6498 record data frame of daily weather observations.
}
\details{
Source?????

\itemize{
   \item{Date POSIXct} 
   \item{Max.TemperatureF integer} 
   \item{Mean.TemperatureF integer} 
   \item{Min.TemperatureF integer} 
   \item{Max.Dew.PointF integer} 
   \item{MeanDew.PointF integer} 
   \item{Min.DewpointF integer} 
   \item{Max.Humidity integer} 
   \item{Mean.Humidity integer} 
   \item{Min.Humidity integer} 
   \item{Max.Sea.Level.PressureIn numeric} 
   \item{Mean.Sea.Level.PressureIn numeric} 
   \item{Min.Sea.Level.PressureIn numeric} 
   \item{Max.VisibilityMiles integer} 
   \item{Mean.VisibilityMiles integer} 
   \item{Min.VisibilityMiles integer} 
   \item{Max.Wind.SpeedMPH integer} 
   \item{Mean.Wind.SpeedMPH integer} 
   \item{Max.Gust.SpeedMPH integer} 
   \item{PrecipitationIn numeric} 
   \item{CloudCover integer} 
   \item{Events factor} 
   \item{WindDirDegrees integer} 
   \item{Fog logical} 
}
}
\keyword{datasets}
