% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\name{draw_topbar}
\alias{draw_topbar}
\title{Draw the color bar across the top of the figure underneath the
title}
\usage{
draw_topbar(bar, fx = c(0.1, 0.1), dy = 0.5, interpolate = FALSE)
}
\arguments{
\item{bar}{a color matrix}

\item{fx}{numeric vector of set backs [left, right] for the ends of the
color bar}

\item{dy}{numeric controls the space between the bottom of the bar and the top
plot as a fraction of string height}

\item{interpolate}{logical, if TRUE interpolate the color bar otherwise sample
- see \code{rasterImage} for details}
}
\value{
a vector of [left, right, bottom, top, width, height, cx, cy] bar coords
}
\description{
Draw the color bar across the top of the figure underneath the
title
}

