% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze.R
\name{is_continuous}
\alias{is_continuous}
\title{Determine if a stsaav object represents a continuous variable}
\usage{
is_continuous(x)
}
\arguments{
\item{x}{a stsaav object}
}
\value{
logical, TRUE if numeric
}
\description{
Determine if a stsaav object represents a continuous variable
}

