% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.stsaav}
\alias{plot.stsaav}
\title{Plot a stsaav object}
\usage{
\method{plot}{stsaav}(x, continuous = is_continuous(x), main = NULL)
}
\arguments{
\item{x}{a stsaav object}

\item{continuous}{logical TRUE to treat vcol as a continuous variable}

\item{main}{character title, if null then not displayed}
}
\description{
Plot a stsaav object
}
