% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_a_anom}
\alias{plot_a_anom}
\title{Plot the seasonal cycle.  Continuous variables are plotted as boxplots while
binary variables are plotted as sums.}
\usage{
plot_a_anom(x, main = "Annual Anomalies")
}
\arguments{
\item{x}{a stsaav object}

\item{main}{the title}
}
\description{
Plot the seasonal cycle.  Continuous variables are plotted as boxplots while
binary variables are plotted as sums.
}
