% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_sample_density}
\alias{plot_sample_density}
\title{Plot the sample density.}
\usage{
plot_sample_density(x, main = pretty_main(x$t_step, what = "Sample Density"))
}
\arguments{
\item{x}{a stsaav object}

\item{main}{the title}
}
\description{
Plot the sample density.
}

