% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_t_anom}
\alias{plot_t_anom}
\title{Plot the seasonal anomaly}
\usage{
plot_t_anom(x, label_format = "\%0.2f", main = pretty_main(x$t_step, what =
  "Anomalies"))
}
\arguments{
\item{x}{a stsaav object}

\item{label_format}{character the formating specification for the color bar}

\item{main}{the title}
}
\description{
Plot the seasonal anomaly
}

