% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_t_cycle}
\alias{plot_t_cycle}
\title{Plot the seasonal cycle.  Continuous variables are plotted as boxplots while
binary variables are plotted as sums.}
\usage{
plot_t_cycle(x, continuous = is_continuous(x), main = pretty_main(x$t_step,
  what = "Cycle"))
}
\arguments{
\item{x}{a stsaav object}

\item{continuous}{logical TRUE to treat vcol as a continuous variable}

\item{main}{character, by default 'Season Cycle'}
}
\description{
Plot the seasonal cycle.  Continuous variables are plotted as boxplots while
binary variables are plotted as sums.
}
