% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_t_ranked_anom}
\alias{plot_t_ranked_anom}
\title{Plot the seasonal anomaly}
\usage{
plot_t_ranked_anom(x, n_sd = c(-3, 3), label_format = "\%0.0f",
  main = pretty_main(x$t_step, what = "Ranked Anomalies"))
}
\arguments{
\item{x}{a stsaav object}

\item{n_sd}{numeric vector, number of standard deviation steps above/below
by default we use -3 to +3}

\item{label_format}{character the formating specification for the color bar}
}
\description{
Plot the seasonal anomaly
}
