% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_t_series}
\alias{plot_t_series}
\title{Plot the timeseries.  Continuous variables are plotted as is while
binary variables are plotted as t_step means.}
\usage{
plot_t_series(x, continuous = is_continuous(x), main = "Time Series")
}
\arguments{
\item{x}{a stsaav object}

\item{continuous}{logical TRUE unless dealing with logical inputs.}

\item{main}{character, the title}
}
\description{
Plot the timeseries.  Continuous variables are plotted as is while
binary variables are plotted as t_step means.
}

