% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{pretty_main}
\alias{pretty_main}
\title{Generate a pretty title.}
\usage{
pretty_main(x, what = c("Anomalies", "Ranked Anomalies", "Cycle")[1])
}
\arguments{
\item{x}{either character (Day, Month or Year) or an stsaav class object}

\item{character, }{by default 'Anomalies' but could be anything}
}
\value{
character such as 'Daily Anomalies'
}
\description{
Generate a pretty title.
}
