% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{show_example}
\alias{show_example}
\title{Show an example output using the built-in dataset}
\usage{
show_example(x = airport_wx, t_step = "Month", vcol = "Fog")
}
\arguments{
\item{x}{a data frame}

\item{t_step}{the time step to analyze}

\item{vcol}{the name of the variable column to analyze}
}
\value{
a stsaav object
}
\description{
Show an example output using the built-in dataset
}
