% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsaav-package.R
\docType{package}
\name{stsaav-package}
\alias{stsaav-package}
\title{Standard Time Series Analysis and Visualization.}
\description{
Simplified analysis and display of standardizedtime series string modeled
after examples found here 
\url{http://www.st.nmfs.noaa.gov/plankton/time-series-methods/index.html}
}
