% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze.R
\name{stsaav}
\alias{stsaav}
\title{Construct a stsaav class object from a data frame.}
\usage{
stsaav(x, what = c("Month", "Week", "Day")[1], tcol = "Date",
  vcol = "value")
}
\arguments{
\item{x}{a data.frame}

\item{what}{the name of the time interval for analysis}

\item{tcol}{character, the name of the time column}

\item{vcol}{character the name of the value column}
}
\value{
a stsaav class object which contains
\itemize{
   \item{data the original data frame with Year, Month, Week and Day columns added}
   \item{xycol charcacter vector of tcol and vcol names}
   \item{ann_sum a vector of annual sums}
   \item{ann_mean a vector of annual means}
   \item{ann_dep a vector of annual deparature from the long term annual mean}
   \item{ann_rdep a vector of annual standardized departures as sd intervals}
   \item{t_step character, the name of the time step (Month, Week, Day)}
   \item{t_x vector of years}
   \item{t_y vector of time steps}
   \item{t_sum matrix of the sum of the 'value' by t_step and year}
   \item{t_n matrix of the observation counts}
   \item{t_mean matrix of the mean observations of 'value'}
   \item{t_sd matrix of the standard deviation of observations of 'value'}
   \item{t_dep matrix of the departure from mean for each time step each year}
   \item{t_rdep matrix of standardized departure from mean for each time step each year}
 }
}
\description{
Construct a stsaav class object from a data frame.
}

