% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stsaav-package.R
\docType{data}
\name{airport_wx}
\alias{airport_wx}
\title{Airport weather data at Narsarsuaq, Greenland}
\format{\preformatted{'data.frame':	6498 obs. of  24 variables:
 $ Date                     : POSIXct, format: "1996-06-30" "1996-07-01" "1996-07-02" ...
 $ Max.TemperatureF         : int  41 60 64 62 62 64 51 59 55 53 ...
 $ Mean.TemperatureF        : int  41 48 55 52 52 53 48 50 50 48 ...
 $ Min.TemperatureF         : int  39 37 46 42 42 42 44 42 46 44 ...
 $ Max.Dew.PointF           : int  36 37 39 41 43 43 46 46 48 46 ...
 $ MeanDew.PointF           : int  35 33 24 38 39 39 41 39 37 44 ...
 $ Min.DewpointF            : int  34 27 16 34 34 32 37 34 23 41 ...
 $ Max.Humidity             : int  81 100 71 81 81 93 100 100 100 100 ...
 $ Mean.Humidity            : int  80 65 32 60 61 62 82 65 61 83 ...
 $ Min.Humidity             : int  76 27 18 34 34 32 62 39 29 67 ...
 $ Max.Sea.Level.PressureIn : num  29.8 29.8 29.9 29.9 29.9 ...
 $ Mean.Sea.Level.PressureIn: num  29.8 29.8 29.8 29.9 29.9 ...
 $ Min.Sea.Level.PressureIn : num  29.8 29.8 29.8 29.9 29.9 ...
 $ Max.VisibilityMiles      : int  6 6 6 6 6 6 6 6 6 6 ...
 $ Mean.VisibilityMiles     : int  6 6 6 6 6 6 6 6 6 6 ...
 $ Min.VisibilityMiles      : int  6 6 6 6 6 6 6 6 6 6 ...
 $ Max.Wind.SpeedMPH        : int  4 12 18 16 13 15 16 15 23 15 ...
 $ Mean.Wind.SpeedMPH       : int  2 4 10 5 5 6 5 6 11 7 ...
 $ Max.Gust.SpeedMPH        : int  NA NA NA NA NA NA NA NA NA NA ...
 $ PrecipitationIn          : num  0 0 0 0 0 0 0 0 0 0 ...
 $ CloudCover               : int  3 2 1 3 1 2 7 5 6 6 ...
 $ Events                   : chr  "" "" "" "" ...
 $ WindDirDegrees           : int  128 258 254 245 290 264 259 245 79 237 ...
 $ Fog                      : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
}}
\usage{
airport_wx
}
\description{
A 24 variable 6498 record data frame of daily weather observations.
}
\details{
Source?????

\itemize{
   \item{Date POSIXct}
   \item{Max.TemperatureF integer}
   \item{Mean.TemperatureF integer}
   \item{Min.TemperatureF integer}
   \item{Max.Dew.PointF integer}
   \item{MeanDew.PointF integer}
   \item{Min.DewpointF integer}
   \item{Max.Humidity integer}
   \item{Mean.Humidity integer}
   \item{Min.Humidity integer}
   \item{Max.Sea.Level.PressureIn numeric}
   \item{Mean.Sea.Level.PressureIn numeric}
   \item{Min.Sea.Level.PressureIn numeric}
   \item{Max.VisibilityMiles integer}
   \item{Mean.VisibilityMiles integer}
   \item{Min.VisibilityMiles integer}
   \item{Max.Wind.SpeedMPH integer}
   \item{Mean.Wind.SpeedMPH integer}
   \item{Max.Gust.SpeedMPH integer}
   \item{PrecipitationIn numeric}
   \item{CloudCover integer}
   \item{Events factor}
   \item{WindDirDegrees integer}
   \item{Fog logical}
}
}
\keyword{datasets}

