% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_sample_density}
\alias{plot_sample_density}
\title{Plot the sample density.}
\usage{
plot_sample_density(x, label_format = "\%0.0f", main = pretty_main(x$t_step,
  what = "Sample Density"))
}
\arguments{
\item{x}{a stsaav object}

\item{label_format}{character the formating specification for the color bar}

\item{main}{the title}
}
\description{
Plot the sample density.
}
