% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_t_anom}
\alias{plot_t_anom}
\title{Plot the seasonal anomaly}
\usage{
plot_t_anom(x, symmetric = TRUE, label_format = "\%0.2f",
  main = pretty_main(x$t_step, what = "Anomalies"))
}
\arguments{
\item{x}{a stsaav object}

\item{symmetric}{logical, if TRUE then force the coloring to be symmetric}

\item{label_format}{character the formating specification for the color bar}

\item{main}{the title}
}
\description{
Plot the seasonal anomaly
}
